SoundIo: Don't compare int with unsigned int

+++ Libraries/SoundIo/SoundIo.c
@@ -206,7 +206,7 @@ struct SoundIo *soundio_create(void) {
 int soundio_connect(struct SoundIo *soundio) {
     int err = 0;
 
-    for (int i = 0; i < ARRAY_LENGTH(available_backends); i += 1) {
+    for (size_t i = 0; i < ARRAY_LENGTH(available_backends); i += 1) {
         enum SoundIoBackend backend = available_backends[i];
         err = soundio_connect_backend(soundio, backend);
         if (!err)
